chore(refactor): init forge module
