+++ src/lib.rs
@@ -10,8 +10,9 @@
 //! - [`input`] models a flake input and its [`input::Follows`] declarations.
 //! - [`lock`] parses `flake.lock` for revs, follows targets, and nested input
 //!   discovery via [`lock::FlakeLock::nested_inputs`].
-//! - [`update`] handles version pins, channel updates, and remote tag lookup
-//!   (paired with [`api`], [`channel`], [`version`]).
+//! - [`forge`] talks to GitHub / Gitea / Forgejo, normalizes versions, and
+//!   applies pin/unpin updates (`forge::api`, `forge::channel`,
+//!   `forge::version`, `forge::update`).
 //! - [`config`] loads `flake-edit.toml`.
 //! - [`cache`] persists URI completion state.
 //! - [`validate`] runs pre-edit lint passes. [`error::FlakeEditError`] is the
@@ -19,12 +20,10 @@
 //!
 //! Feature flags: `tui` enables [`app`] and [`tui`], `diff` enables [`diff`].
 
-pub mod api;
 #[cfg(feature = "tui")]
 pub mod app;
 pub mod cache;
 pub mod change;
-pub mod channel;
 pub mod cli;
 pub mod config;
 #[cfg(feature = "diff")]
@@ -32,12 +31,11 @@ pub mod diff;
 pub mod edit;
 pub mod error;
 pub mod follows;
+pub mod forge;
 pub mod input;
 pub mod lock;
 #[cfg(feature = "tui")]
 pub mod tui;
-pub mod update;
 pub mod uri;
 pub mod validate;
-pub mod version;
 pub mod walk;
